Migrate affiliate.js (#144)
* Creates affiliate.js

Adds a functionality for icarus to immediately after a message is sent say use this fancy link that gives ldsg money!

* Switch affiliate.js to be in ldsg.js

Fixes some small things that Bobby told me to. Should work

* Fixes link naming

Makes it so that the link contains the text "Cdkeys affiliate" or ect

* Update modules/ldsg.js

Bobby asked for some changes. So they are now like that. Cause that makes more sense in whitespace

Co-authored-by: BobbyTheCatfish <46359040+BobbyTheCatfish@users.noreply.github.com>

* Fix Whitespace

* Fixes whitespace

* Update modules/ldsg.js

Fixes some jsdoc things

Co-authored-by: BobbyTheCatfish <46359040+BobbyTheCatfish@users.noreply.github.com>

* Changes from msg.channel.send to msg.reply

Co-authored-by: BobbyTheCatfish <46359040+BobbyTheCatfish@users.noreply.github.com>

* Makes it so that eslint doesn't hate us

Co-authored-by: BobbyTheCatfish <46359040+BobbyTheCatfish@users.noreply.github.com>

---------

Co-authored-by: BobbyTheCatfish <46359040+BobbyTheCatfish@users.noreply.github.com>